Reduce number of files in alerts directory
